✨  added logged-in state
