+++ package.json
@@ -27,6 +27,7 @@
   },
   "homepage": "https://github.com/episodehunter/web#readme",
   "devDependencies": {
+    "@types/history": "^4.6.2",
     "@types/jest": "^22.1.4",
     "@types/react": "^16.0.40",
     "@types/react-dom": "^16.0.4",
@@ -44,6 +45,7 @@
   "dependencies": {
     "@types/auth0-lock": "^10.16.3",
     "auth0-lock": "^11.3.1",
+    "history": "^4.7.2",
     "mobx": "^3.6.1",
     "mobx-react": "^4.4.2",
     "react": "^16.2.0",
