+++ src/history.ts
@@ -1,0 +1,5 @@
+import createHistory from 'history/createBrowserHistory'
+
+export const history = createHistory({
+  forceRefresh: true
+})
