+++ src/index.tsx
@@ -1,19 +1,20 @@
 import * as React from 'react'
 import * as ReactDOM from 'react-dom'
 import styled from 'styled-components'
-import { BrowserRouter as Router, Route } from 'react-router-dom'
+import { Router, Route } from 'react-router-dom'
 import { Provider } from 'mobx-react'
+import { history } from './history'
 import './styles/global'
 import { store } from './store/store'
 import { HomePage } from './pages/home'
-import { SecretPage } from './pages/secret'
+import { LoginPage } from './pages/login'
 
 const App = () => (
   <Provider {...store}>
-    <Router>
+    <Router history={history}>
       <Wrapper>
         <Route exact path="/" component={HomePage} />
-        <Route path="/secret" component={SecretPage} />
+        <Route path="/login" component={LoginPage} />
       </Wrapper>
     </Router>
   </Provider>
